separate send logic from GdplabsChatUI.tsx
